+++ lib/src/models/models.dart
@@ -1,1 +1,2 @@
+export 'source_model.dart';
 export 'weather_model.dart';